Pipe: TsFile batch adds memory control mechanism
